fix incomplete merge
