+++ tests/unit/source_code/test_files.py
@@ -2,6 +2,7 @@ import tempfile
 from pathlib import Path
 from unittest.mock import Mock, create_autospec
 
+import pytest
 from databricks.sdk.service.workspace import Language
 
 from databricks.labs.ucx.source_code.files import Files
@@ -55,7 +56,7 @@ def test_files_supported_language_with_fixer():
     languages.linter(Language.PYTHON).lint.return_value = [Mock(code='some-code')]
     languages.fixer(Language.PYTHON, 'some-code').apply.return_value = "Hi there!"
     files = Files(languages)
-    with tempfile.NamedTemporaryFile(mode = "w+t", suffix = ".py") as file:
+    with tempfile.NamedTemporaryFile(mode="w+t", suffix=".py") as file:
         file.writelines(["import tempfile"])
         path = Path(file.name)
         files.apply(path)
@@ -72,10 +73,9 @@ def test_files_walks_directory():
     languages.fixer.assert_called_with(Language.PYTHON, 'some-code')
     assert languages.fixer.call_count > 1
 
-# TODO the below is unmanageably slow when ran locally, so disabling for now
-# until I get clarifications on what could be done to run it faster
-# created GH issue #1127
-# def test_files_for_cli():
-#     ws = workspace_client_mock()
-#     clazz = Files.for_cli(ws)
-#     assert clazz is not None
+
+@pytest.mark.skip("the below is unmanageably slow when ran locally, so disabling for now, created GH issue #1127")
+def test_files_for_cli():
+    ws = workspace_client_mock()
+    clazz = Files.for_cli(ws)
+    assert clazz is not None
